Squashed commit of the following:
commit 329865fdd98927a18acb446cf0c16fbac1cec7e1
Author: ostrkrs <poluosetr.polukaras@gmail.com>
Date:   Thu Jun 25 18:23:47 2026 +0500

    fixes

commit 30a69f02e4d17b4a5018f7832e74c8c7eeaf47ec
Author: ostrkrs <poluosetr.polukaras@gmail.com>
Date:   Thu Jun 25 10:50:33 2026 +0500

    pasted shi from TeeGee

+++ code/modules/hallucination/station_message.dm
@@ -2,8 +2,12 @@
 	abstract_hallucination_parent = /datum/hallucination/station_message
 	random_hallucination_weight = 1
 	hallucination_tier = HALLUCINATION_TIER_RARE
+	var/require_hearing = TRUE
 
 /datum/hallucination/station_message/start()
+	if(require_hearing && HAS_TRAIT(hallucinator, TRAIT_DEAF))
+		return FALSE
+
 	qdel(src) // To be implemented by subtypes, call parent for easy cleanup
 	return TRUE
